core: split crypt filter methods into a separate package
